requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
python-dateutil==2.9.0.post0
